Routes are all set up all but the API call for OAUTH, everything seems to be in order but testing needs to happen. I think the next step for me (joshi) is to create the API docs using Swagger, and then ask James if he can review the code.
